Laid out new design/framework
Removed old HTML and JS version to start a cleaner, more dynamic remake of the LEDs. This version will allow for creating an object to represent sections/slices of the LED strip instead of hardcoding an array. This will be used in the future to create dynamic rectangles representing LEDs and also set the framework for letting multiple people see live changes on the web app to reflect the LEDs.